drm/rockchip: hdmi: fix 4k10 bit display abnormal
avoid out of value range calculate catmdsclk when 4k10bit and set
scdc high tmds clock ratio when mtmdsclock is more than 340000000

+++ drivers/video/drm/rockchip-inno-hdmi-phy.c
@@ -318,7 +318,7 @@ static inline void inno_update_bits(struct inno_hdmi_phy *inno, u8 reg,
 	inno_write(inno, reg, tmp);
 }
 
-static u32 inno_hdmi_phy_get_tmdsclk(struct inno_hdmi_phy *inno, int rate,
+static u32 inno_hdmi_phy_get_tmdsclk(struct inno_hdmi_phy *inno, unsigned long rate,
 				     int format)
 {
 	u32 tmdsclk;
